requirements.txt
anls>=0.0.2
antlr4-python3-runtime==4.11.1
# requirements
apted==1.0.3
backoff==2.2.1
bert_score
bs4==0.0.2
cairosvg
# vlm
colormath>=3.0.0
datasets>=2.16.1
decord>=0.6.0
distance>=0.1.3
drawsvg
editdistance>=0.8.1
einops
evaluate
func-timeout
httpx==0.27.2
huggingface_hub==0.36.0
imageio
ipdb
jieba>=0.42.1
latex2sympy2_extended
levenshtein>=0.27.1
lpips
lxml>=6.0.2
math-verify
matplotlib
networkx
nltk==3.8.1
num2words
numpy
openai
openai-clip
opencv-python
openpyxl
pandas
pdf2image>=1.17.0
Pillow==10.4.0
polygon3>=3.0.9.1
portalocker
protobuf
pycocoevalcap
pydantic==1.10.12
pylatexenc==2.10
pymupdf
qwen_vl_utils
regex==2024.11.6
requests
rich
rouge>=1.0.1
rouge-score>=0.1.2
scikit-image
scikit-learn>=1.5.2
# For UniSVG
sentence_transformers
setuptools
shapely
spacy
sty

# 基础pip
sympy
tabulate
termcolor
tiktoken
timeout-decorator==0.5.0
torch
# For SArena
torchmetrics
torchvision
tqdm
transformers
typing_extensions
validators==0.34.0
volcengine-python-sdk[ark]
xlsxwriter
zss>=1.2.0
